Added ASMITH_TASK_MEMORY_OPTIMISED
Reduces the size of Task to 8 bytes (32-bit mode).

I don't recomend using this option but it exists if you want to optimise the memory usage of the task scheduling system for some reason.

When enabled:
Maximum number of priority levels reduced to 64.
Exceptions are not allowed to be thrown by task execution or wait conditions.
Only 255 schedulers can exist globally at the same time.
Small overhead to work out which scheduler a task is attached to.